refactor: improved file structure for cleaner imports
