+++ src/components/LaunchModal/InitEntryContent/index.ts
@@ -1,0 +1,2 @@
+export { default } from './InitEntryContent';
+export type { Props as InitEntryContentProps } from './InitEntryContent';
