+++ src/components/LaunchModal/InitEstContent/index.ts
@@ -1,0 +1,2 @@
+export { default } from './InitEstContent';
+export type { Props as InitEstContentProps } from './InitEstContent';
