+++ src/components/LaunchModal/LaunchModal.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Modal } from "@mui/material";
-import InitEntryContent from "./InitEntryContent";
-import InitEstContent from "./InitEstContent";
+import InitEntryContent from "./InitEntryContent/";
+import InitEstContent from "./InitEstContent/";
 
 export type OnChangeEventType = {
   target: {
