+++ src/components/LaunchModal/TimeSelect/index.ts
@@ -1,0 +1,2 @@
+export { default } from './TimeSelect';
+export type { Props as TimeSelectProps } from './TimeSelect';
