+++ src/components/LaunchModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './LaunchModal';
