+++ src/components/Main/TempWarningModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TempWarningModal';
