+++ src/components/Main/TimeTempChart/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TimeTempChart';
