+++ src/components/Main/TimeTempEntry/TimeTempEntry.tsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import styled from "@emotion/styled";
 import { Button, MenuItem, Select, TextField } from "@mui/material";
-import { addTimeTemp } from "../../redux/slices/sessionSlice";
-import { RootState } from "../../redux/store";
-import { getDefaultStartDate, getFormattedTime, getTimeIntervals } from "../../utils";
+import { addTimeTemp } from "../../../redux/slices/sessionSlice";
+import { RootState } from "../../../redux/store";
+import { getDefaultStartDate, getFormattedTime, getTimeIntervals } from "../../../utils";
 
 export type OnCoalsChangeType = React.ChangeEvent<{
   name?: string | undefined;
