+++ src/components/Main/TimeTempTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TimeTempTable';
